+++ exporter/listener_test.go
@@ -6,9 +6,10 @@ import (
 	"testing"
 	"time"
 
+	"log/slog"
+
 	"github.com/alicebob/miniredis/v2"
 	"github.com/redis/go-redis/v9"
-	"golang.org/x/exp/slog"
 )
 
 type mockHandler struct {